Add sequential tapping timeout
Added configurable interval for sequential tapping on the toolbar's
next and previous buttons. Sustained sequential tapping can bring
down Safari on the iPad for galleries with a large number of images
if every tapping event is fired - probably due to them being fired
without enough time for the event processing.