Tidying mouse and touch event handling
Fixes issues caused by Photoswipe jumping to the conclusion that mouse
or touch can be deactivated depending on what is used first.
This conclusion is not good on touch enabled laptops using browsers
which only generate touch events when touching and mouse events when
using the mouse, like Chrome and Opera 15+ on Win 8.
Also fixes an error in the deactivation of the touchstart handler in
onMouseDown, which can ultimately disable both mouse and touch in Chrome
and Opera 15+.