Adding a custom image pid option to the history module
The method for specifying the custom pid to display in
the url hash can be supplied via the `getCurrentImageUID`
option. By default the function will return:

   return _currentItemIndex + 1;

The method for specifying the current image index can be
set using the `getCurrentImageIndex` option. The method
receives the current items array and should return the
numeric index of the current item. By default it returns
the numeric value of the PID hash parameter.