New galleryPIDs option
The getItemAt function may now accept a custom PID and will return the corresponding item whose "pid" property matches the parameter if the "galleryPIDs" option is specified when creating the gallery.  Even if the option is specified, this function still needs to fall back to returning the item with the given index for other components which call it, which it does.  But if there is an index/pid collision, that might cause problems.  So the use of strictly numerical PIDs that might intersect the array indices should probably be dissuaded.