Custom PID support
Custom PID aliases may be provided in the items passed when the gallery is instantiated (similar to the `galleryUID` option) and will be carried forward into the history hash accordingly when the `galleryPIDs` flag is set in options. For example:

options = {
	galleryUID: 78,
	galleryPIDs: true
};

item = {
	src: linkEl.getAttribute('href'),
	w: parseInt(size[0], 10),
	h: parseInt(size[1], 10),
	pid: 'photo1234'
};

will cause the location hash to be `&pid=photo1234&gid=78` when navigating to that item.