Bugfix: Placeholder now loads src rather than msrc
Looks like there was just a typo with this, once the placeholder is loaded it tries to insert it by using the msrc, rather than the src attribute.

I searched for msrc in the project and found nothing, so presumed a typo/refactor mistake.

Have rectified this and it now works correctly. When I use a getThumbBoundsFn the preloaded image is no longer blank.