Fix closing when thumbnail bounds are unknown
Currently, when getThumbBoundsFn is false and the user closes the
gallery, the zoom level and image position is reset. This causes a
visible flash of the initial image if:
* The image is zoomed in
* The user drags down or pinches to close

This changes the closing animation so that only the template is faded
out, keeping the image in the correct place / zoom level.